requirements.txt
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudiowpatch>=0.2.12.7; sys_platform == 'win32'
PyAudio>=0.2.11; sys_platform != 'win32'
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
faster-whisper>=0.10.0
torch>=2.0.0
torchaudio>=2.2.0
speechbrain>=0.5.16
scikit-learn>=1.3.0
transformers>=4.30.0
requests>=2.31.0
python-dotenv>=1.0.0
